Fixed delete function in generic question type
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@17200 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/TestQuestionPool/classes/class.assQuestion.php
@@ -1284,10 +1284,13 @@ class assQuestion
 	{
 		global $ilDB;
 		$additional_table_name = $this->getAdditionalTableName();
-		$query = sprintf("DELETE FROM $additional_table_name WHERE question_fi = %s",
-			$ilDB->quote($question_id . "")
-		);
-		$result = $ilDB->query($query);
+		if (strlen($additional_table_name))
+		{
+			$query = sprintf("DELETE FROM $additional_table_name WHERE question_fi = %s",
+				$ilDB->quote($question_id . "")
+			);
+			$result = $ilDB->query($query);
+		}
 	}
 
 	/**
